Altered Addisjon.java to use floats instead of ints (Week 34, Task 1)

+++ dape1400/week35/Addisjon.java
@@ -7,9 +7,9 @@ public class Addisjon
       String førsteSifferstreng;
       String andreSifferstreng;
 
-      int førstetall;
-      int andretall;
-      int sum;
+      float førstetall;
+      float andretall;
+      float sum;
 
       // Leser inn første tall fra bruker i form av en sifferstreng:
       førsteSifferstreng = JOptionPane.showInputDialog( 
@@ -20,8 +20,8 @@ public class Addisjon
          JOptionPane.showInputDialog( "Skriv inn andre hele tall" );
 
       // Konverterer sifferstrengene til tallverdier:
-      førstetall = Integer.parseInt( førsteSifferstreng );
-      andretall = Integer.parseInt( andreSifferstreng );
+      førstetall = Float.parseFloat( førsteSifferstreng );
+      andretall = Float.parseFloat( andreSifferstreng );
 
       // Adderer tallene og lagrer summen i variabelen sum:
       sum = førstetall + andretall;
